# server/openapi.yaml
openapi: 3.0.0
info:
  title: DenseCore API
  description: High-performance CPU inference engine for LLMs, compatible with OpenAI API.
  version: 2.0.0
  contact:
    name: DenseCore Team
    url: https://github.com/Jake-Network/DenseCore
servers:
  - url: http://localhost:8080/v1
    description: Local development server
paths:
...
  /models/load:
    post:
      summary: Load a model
      operationId: loadModel
      tags:
        - Models
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - model_path
              properties:
                model_path:
                  type: string
                  description: Path to the GGUF model file
                draft_model_path:
                  type: string
                  description: Path to draft model for speculative decoding
                threads:
                  type: integer
                  description: Number of threads to use
      responses:
        '200':
          description: Model loaded successfully
...